+++ packages/polyglot/src/Inference/Drivers/GeminiOAI/GeminiOAIUsageFormat.php
@@ -10,8 +10,8 @@ class GeminiOAIUsageFormat implements CanMapUsage
     #[\Override]
     public function fromData(array $data): Usage {
         return new Usage(
-            inputTokens: $data['usage']['prompt_tokens'] ?? 0,
-            outputTokens: $data['usage']['completion_tokens'] ?? 0,
+            inputTokens: (int) ($data['usage']['prompt_tokens'] ?? 0),
+            outputTokens: (int) ($data['usage']['completion_tokens'] ?? 0),
             cacheWriteTokens: 0,
             cacheReadTokens: 0,
             reasoningTokens: 0,
