+++ packages/polyglot/src/Inference/Drivers/Groq/GroqUsageFormat.php
@@ -10,8 +10,8 @@ class GroqUsageFormat implements CanMapUsage
     #[\Override]
     public function fromData(array $data): Usage {
         return new Usage(
-            inputTokens: $data['x_groq']['usage']['prompt_tokens'] ?? $data['usage']['prompt_tokens'] ?? 0,
-            outputTokens: $data['x_groq']['usage']['completion_tokens'] ?? $data['usage']['completion_tokens'] ?? 0,
+            inputTokens: (int) ($data['x_groq']['usage']['prompt_tokens'] ?? $data['usage']['prompt_tokens'] ?? 0),
+            outputTokens: (int) ($data['x_groq']['usage']['completion_tokens'] ?? $data['usage']['completion_tokens'] ?? 0),
             cacheWriteTokens: 0,
             cacheReadTokens: 0,
             reasoningTokens: 0,
