+++ packages/polyglot/src/Inference/Drivers/OpenAI/OpenAIUsageFormat.php
@@ -10,11 +10,11 @@ class OpenAIUsageFormat implements CanMapUsage
     #[\Override]
     public function fromData(array $data): Usage {
         return new Usage(
-            inputTokens: $data['usage']['prompt_tokens'] ?? 0,
-            outputTokens: $data['usage']['completion_tokens'] ?? 0,
+            inputTokens: (int) ($data['usage']['prompt_tokens'] ?? 0),
+            outputTokens: (int) ($data['usage']['completion_tokens'] ?? 0),
             cacheWriteTokens: 0,
-            cacheReadTokens: $data['usage']['prompt_tokens_details']['cached_tokens'] ?? 0,
-            reasoningTokens: $data['usage']['prompt_tokens_details']['reasoning_tokens'] ?? 0,
+            cacheReadTokens: (int) ($data['usage']['prompt_tokens_details']['cached_tokens'] ?? 0),
+            reasoningTokens: (int) ($data['usage']['prompt_tokens_details']['reasoning_tokens'] ?? 0),
         );
     }
 }
